Rename and refactor the Core API test manager
